Adds logic that allows a user to delete own ads
- Adds logic that allows user to update an ad
